crawler/pyproject.toml
[project]
name = "lottery-crawler"
version = "0.1.0"
description = "Thai lottery statistics crawler using crawl4ai"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
lottery-crawler = "lottery_crawler.__main__:main"
